+++ R/output.R
@@ -10,5 +10,9 @@ write_stderr <- function(text) {
 
 remap_output_streams <- function() {
   output <- import("rpytools.output")
-  output$remap_output_streams(write_stdout, write_stderr)
+  output$remap_output_streams(
+    write_stdout, 
+    write_stderr,
+    tty = interactive() || isatty(stdout())
+  )
 }